+++ lib/src/translate_options/translate_options.dart
@@ -8,7 +8,11 @@ import 'package:file/file.dart';
 enum ModelProvider {
   gemini('gemini', 'Gemini'),
   vertexAi('vertex-ai', 'Vertex AI'),
-  openAi('open-ai', 'Open AI');
+  openAi('open-ai', 'Open AI'),
+  customOpenAiCompatible(
+    'custom',
+    'Custom Open AI compatible',
+  );
 
   const ModelProvider(this.key, this.name);
 
@@ -56,13 +60,16 @@ class TranslateOptions {
   const TranslateOptions({
     required this.modelProvider,
     required this.model,
+    required this.customModel,
     required this.apiKey,
     required this.vertexAiProjectUrl,
+    required this.customModelProviderBaseUrl,
     required bool? disableSafety,
     required this.context,
     required this.arbDir,
     required String? templateArbFile,
     required this.excludeLocales,
+    required this.batchSize,
     required bool? useEscaping,
     required bool? relaxSyntax,
   })  : disableSafety = disableSafety ?? false,
@@ -79,13 +86,16 @@ class TranslateOptions {
 
   final ModelProvider modelProvider;
   final Model model;
+  final String? customModel;
   final String apiKey;
   final Uri? vertexAiProjectUrl;
+  final Uri? customModelProviderBaseUrl;
   final bool disableSafety;
   final String? context;
   final String arbDir;
   final String templateArbFile;
   final List<String>? excludeLocales;
+  final int batchSize;
   final bool useEscaping;
   final bool relaxSyntax;
 
@@ -107,14 +117,27 @@ class TranslateOptions {
     final modelProvider = argResults.modelProvider ??
         yamlResults.modelProvider ??
         ModelProvider.gemini;
+
     final model = argResults.model ??
         yamlResults.model ??
         (modelProvider == ModelProvider.openAi
             ? Model.gpt35Turbo
             : Model.gemini10Pro);
+    final customModel = argResults.customModel ?? yamlResults.customModel;
 
-    if (!model.providers.contains(modelProvider)) {
-      throw ModelProviderMismatchException();
+    if (modelProvider != ModelProvider.customOpenAiCompatible) {
+      if (!model.providers.contains(modelProvider)) {
+        throw ModelProviderMismatchException();
+      }
+    } else {
+      if (customModel == null) {
+        throw MissingCustomModelException();
+      }
+    }
+
+    if (modelProvider == ModelProvider.customOpenAiCompatible &&
+        customModel == null) {
+      throw MissingCustomModelException();
     }
 
     final vertexAiProjectUrlString =
@@ -135,6 +158,24 @@ class TranslateOptions {
       }
     }
 
+    final customModelProviderBaseUrlString =
+        argResults.customModelProviderBaseUrl ??
+            yamlResults.customModelProviderBaseUrl;
+    final Uri? customModelProviderBaseUrl =
+        customModelProviderBaseUrlString != null
+            ? Uri.tryParse(customModelProviderBaseUrlString)
+            : null;
+
+    if (modelProvider == ModelProvider.customOpenAiCompatible) {
+      if (customModelProviderBaseUrlString == null) {
+        throw MissingCustomModelProviderBaseUrlException();
+      }
+
+      if (customModelProviderBaseUrl == null) {
+        throw InvalidCustomModelProviderBaseUrlException();
+      }
+    }
+
     final context = argResults.context ?? yamlResults.context;
 
     if (context != null && context.length > maxContextLength) {
@@ -143,7 +184,9 @@ class TranslateOptions {
 
     return TranslateOptions(
       modelProvider: modelProvider,
+      customModelProviderBaseUrl: customModelProviderBaseUrl,
       model: model,
+      customModel: customModel,
       apiKey: apiKey,
       vertexAiProjectUrl: vertexAiProjectUrl,
       disableSafety: argResults.disableSafety ?? yamlResults.disableSafety,
@@ -154,6 +197,7 @@ class TranslateOptions {
       templateArbFile:
           argResults.templateArbFile ?? yamlResults.templateArbFile,
       excludeLocales: argResults.excludeLocales ?? yamlResults.excludeLocales,
+      batchSize: argResults.batchSize ?? yamlResults.batchSize ?? 4096,
       useEscaping: argResults.useEscaping ?? yamlResults.useEscaping,
       relaxSyntax: argResults.relaxSyntax ?? yamlResults.relaxSyntax,
     );